requirements.txt
fastapi
uvicorn
pandas
google-api-python-client
google-auth
jinja2
python-multipart